+++ src/widgets/notesubfoldertree.h
@@ -1,12 +1,15 @@
 #pragma once
 
 #include <QTreeWidget>
+#include <QVector>
 
 class NoteSubFolder;
 class Note;
+class QMenu;
 
 class NoteSubFolderTree : public QTreeWidget {
     Q_OBJECT
+
    public:
     NoteSubFolderTree(QWidget *parent = nullptr);
 
@@ -21,6 +24,8 @@ class NoteSubFolderTree : public QTreeWidget {
 
     static QMenu *contextMenu(QTreeWidget *parent);
     static void removeSelectedNoteSubFolders(QTreeWidget *treeWidget);
+    static void moveSelectedNoteSubFoldersToParent(QTreeWidget *treeWidget,
+                                                   int destinationParentId);
 
    public Q_SLOTS:
     void onItemExpanded(QTreeWidgetItem *item);
@@ -39,6 +44,9 @@ class NoteSubFolderTree : public QTreeWidget {
     void initConnections();
     QTreeWidgetItem *addNoteSubFolder(QTreeWidgetItem *parentItem,
                                       const NoteSubFolder &noteSubFolder);
+    static QMenu *buildMoveDestinationMenuTree(QTreeWidget *treeWidget, QMenu *parentMenu,
+                                               int parentNoteSubFolderId,
+                                               const QVector<int> &forbiddenDestinationIds);
 
    protected:
     bool eventFilter(QObject *obj, QEvent *event) override;